Convert Listener-Store to TS
